% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_row}
\alias{create_row}
\title{Create Row}
\usage{
create_row(input_file)
}
\arguments{
\item{input_file}{A PDF file location to be scraped and converted.}
}
\value{
A data frame representing structured data for the input file.
}
\description{
Create Row
}
\examples{
\dontrun{
row <- create_row('~/courses/API-101B.pdf')
}
}
