% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_course_number}
\alias{get_course_number}
\title{Get Course Number}
\usage{
get_course_number(lines)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}
}
\value{
A seven- or eight-character course identifier.
}
\description{
Get Course Number
}
\examples{
\dontrun{
course_number <- get_course_number(lines)
}
}
