% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_eval_format}
\alias{get_eval_format}
\title{Get Evaluation Format}
\usage{
get_eval_format(lines)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}
}
\value{
Designation as "Old" or "New" depending on semester and year
}
\description{
Get Evaluation Format
}
\examples{
\dontrun{
eval_format <- get_eval_format(lines)
}
}
