% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_faculty_name}
\alias{get_faculty_name}
\title{Get Faculty Name}
\usage{
get_faculty_name(lines)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}
}
\value{
The name of the faculty member(s) teaching the course. If multiple
faculty members teach the course, their names will be separated with a
backslash.
}
\description{
Get Faculty Name
}
\examples{
\dontrun{
faculty_name <- get_faculty_name(lines)
}
}
