% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_file_paths}
\alias{get_file_paths}
\title{Get File Paths}
\usage{
get_file_paths(list_of_directories)
}
\arguments{
\item{list_of_directories}{Single directory or list/vector of target directories.}
}
\value{
Vector containing full path to each file in each target directory regardless of working directory.
}
\description{
Get File Paths
}
\examples{
\dontrun{
desktop <- get_file_paths('~/Desktop')
}
\dontrun{
desktop_plus <- get_file_paths(c('~/', '~/Desktop'))
}
}
