% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_line_number}
\alias{get_line_number}
\title{Get Line Number}
\usage{
get_line_number(lines = lines, question_number)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}

\item{question_number}{The question number (1-19) to be identified.}
}
\value{
A number representing the list element in which the question is found.
}
\description{
Get Line Number
}
\examples{
\dontrun{
line_number <- get_line_number(lines, 01)
}
}
