% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_question}
\alias{get_question}
\title{Get Question}
\usage{
get_question(lines = lines, question_number)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}

\item{question_number}{The question number (1-19) for which a mean response is desired.}
}
\value{
The mean response from all respondents for the specified question.
}
\description{
Get Question
}
\examples{
\dontrun{
question_01 <- get_question(lines, 01)
}
}
