% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_respondents}
\alias{get_respondents}
\title{Get Respondents}
\usage{
get_respondents(lines)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}
}
\value{
The number of respondents (HKS plus non-HKS) who submitted a course evaluation.
}
\description{
Get Respondents
}
\examples{
\dontrun{
respondents <- get_respondents(lines)
}
}
