% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_semester}
\alias{get_semester}
\title{Get Semester}
\usage{
get_semester(lines)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}
}
\value{
The semester and year in which the course was taught.
}
\description{
Get Semester
}
\examples{
\dontrun{
semester <- get_semester(lines)
}
}
