% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_year}
\alias{get_year}
\title{Get Year}
\usage{
get_year(lines)
}
\arguments{
\item{lines}{Input file represented as list object (output of \code{import_lines}).}
}
\value{
The year in which the course was taught.
}
\description{
Get Year
}
\examples{
\dontrun{
year <- get_year(lines)
}
}
