% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hksreviewr-package.R
\docType{package}
\name{hksreviewr-package}
\alias{hksreviewr}
\alias{hksreviewr-package}
\title{hksreviewr: Scrape HKS Course Evaluation PDFs}
\description{
This package contains functions to create a
    data frame from HKS course evaluations in PDF format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rbrellen/hksreviewr}
  \item Report bugs at \url{https://github.com/rbrellen/hksreviewr/issues}
}

}
\keyword{internal}
