% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{import_lines}
\alias{import_lines}
\title{Import Lines}
\usage{
import_lines(input_file)
}
\arguments{
\item{input_file}{A PDF file location to be scraped and converted.}
}
\value{
List object with each element representing a line of the source PDF.
}
\description{
Import Lines
}
\examples{
\dontrun{
lines <- import_lines('~/courses/API-101B.pdf')
}
}
